*: ready for the 3rd submission

+++ backend/src/main/java/com/bugenzhao/bookstore_backend/service/BookService.java
@@ -3,10 +3,10 @@ package com.bugenzhao.bookstore_backend.service;
 import java.util.List;
 import java.util.Optional;
 
-import com.bugenzhao.bookstore_backend.entity.Book;
+import com.bugenzhao.bookstore_backend.entity.db.Book;
 
 public interface BookService {
-    Optional<Book> findById(int bookId);
+    Optional<Book> findById(long bookId);
 
     List<Book> findAll();
 }
